--- templates/buy-me-a-coffee-app/src/features/BuyMeACoffee/index.ts
@@ -1,2 +1,0 @@
-export * from './FormBuyCoffee';
-export * from './Memos';
