+++ templates/buy-me-a-coffee-app/src/pages/home/components/FormBuyCoffee.tsx
@@ -120,4 +120,4 @@ function FormBuyCoffee({ onComplete }: FormBuyCoffeeProps) {
   );
 }
 
-export { FormBuyCoffee };
+export default FormBuyCoffee;
