+++ templates/buy-me-a-coffee-app/src/pages/home/components/MemoCard.tsx
@@ -35,4 +35,4 @@ function MemoCard({ name, message, timestamp }: CoffeeMemo) {
   );
 }
 
-export { MemoCard };
+export default MemoCard;
